# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pleer Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';

export interface DashboardStats {
  totalListings: number;
  activeListings: number;
  totalUsers: number;
  totalMessages: number;
  sellListings?: number;
  buyListings?: number;
  topCategories?: Category[];
}