# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pleer Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingInputType } from './listingInputType';

export interface ListingInput {
  /** @minLength 3 */
  title: string;
  description?: string;
  type: ListingInputType;
  /** @minimum 0 */
  price: number;
  currency?: string;
  category?: string;
  condition?: string;
  imageUrl?: string;
  city?: string;
}