# lib/api-zod/src/generated/types/listingsPage.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pleer Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from './listing';

export interface ListingsPage {
  items: Listing[];
  total: number;
  page: number;
  limit: number;
}