# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pleer Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  conversationId: number;
  senderId: string;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderAvatarUrl?: string | null;
  text: string;
  createdAt: string;
}